Small cleanup from inspection
git-svn-id: http://myrrix-recommender.googlecode.com/svn/trunk@557 0d189622-dca1-7e23-9e30-e8815eb7bd2f

+++ common/src/net/myrrix/common/random/RandomUtils.java
@@ -228,7 +228,7 @@ public final class RandomUtils {
                                               RandomGenerator random) {
     LongPrimitiveIterator it;
     if (n < streamSize) {
-      it = new SamplingLongPrimitiveIterator(random, stream, (double) n / (double) streamSize);      
+      it = new SamplingLongPrimitiveIterator(random, stream, (double) n / streamSize);      
     } else {
       it = stream;
     }
